feat: remove legacy model and adopt new architecture
