+++ civic_digital_twins/dt_model/__init__.py
@@ -1,7 +1,9 @@
 """The dt_model package implements the tool for digital twins modeling and simulation."""
 
-from .model.model import Model
+from .model.abstract_model import AbstractModel
+from .model.instantiated_model import InstantiatedModel
 from .simulation.ensemble import Ensemble
+from .simulation.evaluation import Evaluation
 from .symbols.constraint import Constraint
 from .symbols.context_variable import (
     CategoricalContextVariable,
@@ -13,15 +15,17 @@ from .symbols.index import ConstIndex, Index, LognormDistIndex, SymIndex, Triang
 from .symbols.presence_variable import PresenceVariable
 
 __all__ = [
+    "AbstractModel",
     "CategoricalContextVariable",
     "Constraint",
     "ConstIndex",
     "ContextVariable",
     "ContinuousContextVariable",
     "Ensemble",
+    "Evaluation",
     "Index",
+    "InstantiatedModel",
     "LognormDistIndex",
-    "Model",
     "PresenceVariable",
     "SymIndex",
     "TriangDistIndex",
